+++ lib/lightning_web/components/viewers.ex
@@ -16,6 +16,7 @@ defmodule LightningWeb.Components.Viewers do
 
   alias Lightning.Invocation.Dataclip
   alias LightningWeb.Components.Icon
+  alias Phoenix.LiveView.JS
 
   require Lightning.Run
 
@@ -68,12 +69,18 @@ defmodule LightningWeb.Components.Viewers do
         </span>
       </div>
     <% else %>
-      <div class="relative flex grow">
-        <.log_level_filter
-          :if={@current_user}
-          id={"#{@id}-filter"}
-          selected_level={@selected_log_level}
-        />
+      <div class="flex flex-col grow rounded-md bg-slate-700 font-mono text-gray-200">
+        <div class="border-b border-slate-500">
+          <div class="mx-auto px-4">
+            <div class="flex h-8 flex-row-reverse items-center">
+              <.log_level_filter
+                :if={@current_user}
+                id={"#{@id}-filter"}
+                selected_level={@selected_log_level}
+              />
+            </div>
+          </div>
+        </div>
 
         <div
           id={@id}
@@ -87,10 +94,7 @@ defmodule LightningWeb.Components.Viewers do
           data-viewer-el={"#{@id}-viewer"}
         >
           <div class="relative grow">
-            <div
-              id={"#{@id}-nothing-yet"}
-              class="relative rounded-md p-12 text-center bg-slate-700 font-mono text-gray-200"
-            >
+            <div id={"#{@id}-nothing-yet"} class="relative p-12 text-center">
               <.text_ping_loader>
                 Nothing yet
               </.text_ping_loader>
@@ -111,77 +115,73 @@ defmodule LightningWeb.Components.Viewers do
     assigns =
       assign(assigns,
         log_levels: [
-          {"debug", "Show all logs"},
-          {"info", "Standard logging, excluding low-level debug logs (default)"},
-          {"warn", "Only show warnings, errors and key logs"},
-          {"error", "Only show errors and key logs"}
+          "debug",
+          "info",
+          "warn",
+          "error"
         ]
       )
 
     ~H"""
-    <div id={@id} class="absolute top-0 right-4 z-50">
-      <.modal
-        id={"#{@id}-modal"}
-        close_on_click_away={false}
-        close_on_keydown={false}
-      >
-        <:title>
-          <div class="flex justify-between gap-2">
-            <span class="font-bold">
-              Configure log levels
+    <div id={@id} class="z-50">
+      <div class="relative">
+        <button
+          type="button"
+          class="grid w-full cursor-pointer grid-cols-1 bg-inherit text-left sm:text-sm/6 hover:bg-slate-600 border-b border-gray-200"
+          aria-haspopup="listbox"
+          aria-expanded="true"
+          phx-click={show_dropdown("#{@id}-dropdown")}
+        >
+          <span class="col-start-1 row-start-1 truncate pr-6">
+            Level: <%= @selected_level %>
+          </span>
+          <.icon
+            name="hero-chevron-down"
+            class="col-start-1 row-start-1 size-5 self-center justify-self-end"
+            aria-hidden="true"
+            data-slot="icon"
+          />
+        </button>
+        <ul
+          id={"#{@id}-dropdown"}
+          class="hidden absolute z-10 mt-1 max-h-60 w-full overflow-auto rounded-md bg-slate-600 py-1 text-base shadow-lg ring-1 ring-black/5 focus:outline-none sm:text-sm"
+          tabindex="-1"
+          role="listbox"
+          phx-click-away={hide_dropdown("#{@id}-dropdown")}
+        >
+          <li
+            :for={log_level <- @log_levels}
+            id={"#{@id}-dropdown-#{log_level}-option"}
+            class="relative cursor-default select-none py-2 pl-8 pr-4 hover:bg-slate-500"
+            role="option"
+            phx-click={
+              JS.push("save-log-filter", value: %{desired_log_level: log_level})
+              |> hide_dropdown("#{@id}-dropdown")
+            }
+          >
+            <span class={[
+              "block truncate",
+              if(log_level == @selected_level,
+                do: "font-semibold",
+                else: "font-normal"
+              )
+            ]}>
+              <%= log_level %>
             </span>
-
-            <button
-              phx-click={hide_modal("#{@id}-modal")}
-              type="button"
-              class="rounded-md bg-white text-gray-400 hover:text-gray-500 focus:outline-none"
-              aria-label={gettext("close")}
+            <span
+              :if={log_level == @selected_level}
+              class="absolute inset-y-0 left-0 flex items-center pl-1.5 text-blue-400"
             >
-              <span class="sr-only">Close</span>
-              <.icon name="hero-x-mark" class="h-5 w-5 stroke-current" />
-            </button>
-          </div>
-        </:title>
-        <.form id={"#{@id}-form"} for={to_form(%{})} phx-change="save-log-filter">
-          <fieldset>
-            <p class="mt-1 text-sm/6 text-gray-600">
-              Which logs do you prefer to see?
-            </p>
-            <div class="mt-6 space-y-4">
-              <%= for {level, description} <- @log_levels do %>
-                <div class="relative flex items-start">
-                  <div class="flex h-6 items-center">
-                    <.input
-                      type="radio"
-                      id={"desired_level_#{level}"}
-                      name="desired_log_level"
-                      value={level}
-                      checked={level == @selected_level}
-                      class="relative size-4 appearance-none rounded-full border border-gray-300 bg-white before:absolute before:inset-1 before:rounded-full before:bg-white checked:border-indigo-600 checked:bg-indigo-600 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600 disabled:border-gray-300 disabled:bg-gray-100 disabled:before:bg-gray-400 forced-colors:appearance-auto forced-colors:before:hidden [&:not(:checked)]:before:hidden"
-                    />
-                  </div>
-                  <div class="ml-3 text-sm/6">
-                    <.label
-                      for={"desired_level_#{level}"}
-                      class="font-medium text-gray-900 uppercase"
-                    >
-                      <%= level %>
-                    </.label>
-                    <p class="text-gray-500"><%= description %></p>
-                  </div>
-                </div>
-              <% end %>
-            </div>
-          </fieldset>
-        </.form>
-      </.modal>
-      <button
-        phx-click={show_modal("#{@id}-modal")}
-        type="button"
-        class="rounded-bl-md rounded-br-md p-1 pt-0 opacity-70 hover:opacity-100 content-center bg-blue-500 text-blue-900"
-      >
-        <.icon name="hero-cog-6-tooth" class="h-4 w-4 inline-block align-middle" />
-      </button>
+              <.icon
+                name="hero-check"
+                class="size-5 font-semibold"
+                aria-hidden="true"
+                data-slot="icon"
+              />
+            </span>
+          </li>
+        </ul>
+      </div>
     </div>
     """
   end